:fire: 实现后台管理页面（CRUD）& JSON统一传输格式 & 全局异常处理 & 跨域问题

+++ 03.SpringBoot/09.BookSystem/src/main/resources/db/schema.sql
@@ -1,6 +1,6 @@
 drop table if exists tb_book;
 create table tb_book (
-  id int auto_increment primary key comment '编号',
+  id bigint auto_increment primary key comment '编号',
   name varchar(64) not null default '' comment '书名',
   author varchar(128) not null default '' comment '作者',
   publisher varchar(128) not null default '' comment '出版社',
